added request

+++ app/src/main/java/com/furkanmulayim/birikio/feature/screen_home/data/model/AllMoneys.kt
@@ -3,11 +3,11 @@ package com.furkanmulayim.birikio.feature.screen_home.data.model
 data class AllMoneys(
     val usd: Double,
     val eur: Double,
-    val grams24: Int,
-    val grams22: Int,
-    val quart: Int,
-    val half: Int,
-    val full: Int,
-    val resat: Int,
+    val grams24: Double,
+    val grams22: Double,
+    val quart: Double,
+    val half: Double,
+    val full: Double,
+    val resat: Double,
 )
